qascribe/backend/pyproject.toml
[project]
name = "qascribe-backend"
version = "2.0.0"
description = "AI-QAScribe backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "celery>=5.4.0",
    "redis>=5.2.0",
    "boto3>=1.35.0",
    "python-multipart>=0.0.18",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "anthropic>=0.40.0",
    "openai>=1.55.0",
    "google-genai>=0.3.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
